+++ examples/coinbase-example/examples/cb-order-sender.rs
@@ -165,7 +165,7 @@ fn main() -> anyhow::Result<()> {
         &app,
         &store_factory,
         &log_factory,
-        ConnectionMode::SingleThreaded,
+        FixSocketServerKind::SingleThreaded,
     )?;
 
     // Start the engine.