+++ quickfix/src/initiator.rs
@@ -8,8 +8,9 @@ use quickfix_ffi::{
 
 use crate::{
     utils::{ffi_code_to_bool, ffi_code_to_result},
-    Application, ApplicationCallback, ConnectionHandler, ConnectionMode, FfiMessageStoreFactory,
-    LogCallback, LogFactory, QuickFixError, Session, SessionContainer, SessionId, SessionSettings,
+    Application, ApplicationCallback, ConnectionHandler, FfiMessageStoreFactory,
+    FixSocketServerKind, LogCallback, LogFactory, QuickFixError, Session, SessionContainer,
+    SessionId, SessionSettings,
 };
 
 /// Socket implementation of establishing connections handler.
@@ -38,7 +39,7 @@ where
         application: &'a Application<A>,
         store_factory: &'a S,
         log_factory: &'a LogFactory<L>,
-        server_mode: ConnectionMode,
+        server_mode: FixSocketServerKind,
     ) -> Result<Self, QuickFixError> {
         match unsafe {
             FixInitiator_new(