+++ quickfix/tests/utils/checker.rs
@@ -40,14 +40,14 @@ pub fn run(
         &app_sender,
         &message_store_factory_sender,
         &log_factory,
-        ConnectionMode::default(),
+        FixSocketServerKind::default(),
     )?;
     let mut socket_receiver = Acceptor::try_new(
         &settings_receiver,
         &app_receiver,
         &message_store_factory_receiver,
         &log_factory,
-        ConnectionMode::default(),
+        FixSocketServerKind::default(),
     )?;
 
     // Start the app